test: ➕ add separate file for tests
