+++ test/index.ts
@@ -40,10 +40,8 @@ import {
   AsyncOrSync,
   Awaited,
   Newable,
-  PickKeys,
   IsTuple,
   Writable,
-  StrictOmit,
   OmitProperties,
   isExact,
   IsUnknown,
@@ -749,57 +747,6 @@ function testBuildable() {
     Assert<IsExact<Buildable<DeepReadonly<ComplexNestedRequired>>, ComplexNestedPartial>>,
     Assert<IsExact<Buildable<ComplexNestedRequired>, ComplexNestedPartial>>,
     Assert<IsExact<Buildable<ComplexNestedReadonly>, ComplexNestedPartial>>,
-  ];
-}
-
-function testStrictOmit() {
-  type cases = [
-    // @ts-expect-error works only with records
-    StrictOmit<number, never>,
-    // @ts-expect-error works only with records
-    StrictOmit<string, never>,
-    // @ts-expect-error works only with records
-    StrictOmit<boolean, never>,
-    // @ts-expect-error works only with records
-    StrictOmit<bigint, never>,
-    // @ts-expect-error works only with records
-    StrictOmit<symbol, never>,
-    // @ts-expect-error works only with records
-    StrictOmit<undefined, never>,
-    // @ts-expect-error works only with records
-    StrictOmit<null, never>,
-    Assert<IsExact<StrictOmit<Function, never> & Function, Function>>,
-    Assert<IsExact<StrictOmit<Date, never> & Date, Date>>,
-    Assert<IsExact<StrictOmit<Error, never> & Error, Error>>,
-    Assert<IsExact<StrictOmit<RegExp, never> & RegExp, RegExp>>,
-    Assert<IsExact<StrictOmit<Map<string, boolean>, never> & Map<string, boolean>, Map<string, boolean>>>,
-    Assert<
-      IsExact<
-        StrictOmit<ReadonlyMap<string, boolean>, never> & ReadonlyMap<string, boolean>,
-        ReadonlyMap<string, boolean>
-      >
-    >,
-    Assert<
-      IsExact<
-        StrictOmit<WeakMap<{ key: string }, boolean>, never> & WeakMap<{ key: string }, boolean>,
-        WeakMap<{ key: string }, boolean>
-      >
-    >,
-    Assert<IsExact<StrictOmit<Set<string>, never> & Set<string>, Set<string>>>,
-    Assert<IsExact<StrictOmit<ReadonlySet<string>, never> & ReadonlySet<string>, ReadonlySet<string>>>,
-    Assert<IsExact<StrictOmit<Promise<number>, never> & Promise<number>, Promise<number>>>,
-    Assert<IsExact<StrictOmit<{}, never>, {}>>,
-    Assert<IsExact<StrictOmit<{ a: 1 }, never>, { a: 1 }>>,
-    Assert<IsExact<StrictOmit<{ a?: 1 }, never>, { a?: 1 }>>,
-    Assert<IsExact<StrictOmit<{ a: 1 }, "a">, {}>>,
-    Assert<IsExact<StrictOmit<{ a?: 1 }, "a">, {}>>,
-    Assert<IsExact<StrictOmit<{ a?: 1 }, "a">, {}>>,
-    // we don't prohibit arrays and tuples, but return never for them
-    Assert<IsExact<StrictOmit<readonly [], never>, never>>,
-    Assert<IsExact<StrictOmit<readonly [1, 2, 3], never>, never>>,
-    Assert<IsExact<StrictOmit<readonly number[], never>, never>>,
-    Assert<IsExact<StrictOmit<Array<number>, never>, never>>,
-    Assert<IsExact<StrictOmit<ReadonlyArray<number>, never>, never>>,
   ];
 }
